requirements.txt
anthropic>=0.40.0
resend>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
tweepy>=4.14.0
praw>=7.7.0
requests>=2.31.0
